requirements.txt
Flask==1.1.2
Flask-MySQL==1.5.2
Flask-RESTful==0.3.8
PyMySQL==1.0.2
python-dotenv==0.19.0
gunicorn
DateTime
pytz
pyzbar
Pillow